requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.16.1
fsspec==2025.3.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
h2==4.1.0
hf-xet==1.2.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.0.1
idna==3.11
Jinja2==3.1.6
jiter==0.9.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
openai==2.2.0
packaging==25.0
pdfminer.six==20250324
pillow==10.4.0
portalocker==2.10.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.24.11
python-dotenv==1.0.1
PyYAML==6.0.3
qdrant-client==1.12.1
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
urllib3==2.2.3
